syscall has sp

+++ e8/types.go
@@ -3,7 +3,7 @@ package e8
 
 // Section types
 const (
-	None uint16 = iota
+	None uint8 = iota
 	Code
 	Data
 	Zeros // a.k.a. BSS
